requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
opencv-python>=4.8.0
fer>=22.4.0
textblob>=0.17.1
spacy>=3.7.2
transformers>=4.35.0
torch>=2.1.0
requests>=2.31.0
python-jose>=3.3.0
plotly>=5.18.0
pandas>=2.1.0
pillow>=10.0.0
mediapipe>=0.10.8
pytest>=7.4.0
python-decouple>=3.8 